+++ flash/core/classification.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Tuple, Union
+from typing import Any, Callable, Dict, List, Mapping, Optional, Sequence, Union
 
 import torch
 import torch.nn.functional as F
@@ -24,7 +24,6 @@ from flash.core.model import Task
 from flash.core.utilities.imports import _FIFTYONE_AVAILABLE
 
 if _FIFTYONE_AVAILABLE:
-    import fiftyone as fo
     from fiftyone.core.labels import Classification, Classifications
 else:
     Classification, Classifications = None, None
@@ -83,34 +82,43 @@ class ClassificationSerializer(Serializer):
         return self._mutli_label
 
 
-class Logits(ClassificationSerializer):
+class PredsClassificationSerializer(ClassificationSerializer):
+    """A :class:`~flash.core.classification.ClassificationSerializer` which gets the
+    :attr:`~flash.core.data.data_source.DefaultDataKeys.PREDS` from the sample.
+    """
+
+    def serialize(self, sample: Any) -> Any:
+        if isinstance(sample, Mapping) and DefaultDataKeys.PREDS in sample:
+            sample = sample[DefaultDataKeys.PREDS]
+        if not isinstance(sample, torch.Tensor):
+            sample = torch.tensor(sample)
+        return sample
+
+
+class Logits(PredsClassificationSerializer):
     """A :class:`.Serializer` which simply converts the model outputs (assumed to be logits) to a list."""
 
     def serialize(self, sample: Any) -> Any:
-        sample = sample[DefaultDataKeys.PREDS] if isinstance(sample, Dict) else sample
-        sample = torch.tensor(sample)
-        return sample.tolist()
+        return super().serialize(sample).tolist()
 
 
-class Probabilities(ClassificationSerializer):
+class Probabilities(PredsClassificationSerializer):
     """A :class:`.Serializer` which applies a softmax to the model outputs (assumed to be logits) and converts to a
     list."""
 
     def serialize(self, sample: Any) -> Any:
-        sample = sample[DefaultDataKeys.PREDS] if isinstance(sample, Dict) else sample
-        sample = torch.tensor(sample)
+        sample = super().serialize(sample)
         if self.multi_label:
             return torch.sigmoid(sample).tolist()
         return torch.softmax(sample, -1).tolist()
 
 
-class Classes(ClassificationSerializer):
+class Classes(PredsClassificationSerializer):
     """A :class:`.Serializer` which applies an argmax to the model outputs (either logits or probabilities) and
     converts to a list.
 
     Args:
         multi_label: If true, treats outputs as multi label logits.
-
         threshold: The threshold to use for multi_label classification.
     """
 
@@ -120,8 +128,7 @@ class Classes(ClassificationSerializer):
         self.threshold = threshold
 
     def serialize(self, sample: Any) -> Union[int, List[int]]:
-        sample = sample[DefaultDataKeys.PREDS] if isinstance(sample, Dict) else sample
-        sample = torch.tensor(sample)
+        sample = super().serialize(sample)
         if self.multi_label:
             one_hot = (sample.sigmoid() > self.threshold).int().tolist()
             result = []
@@ -139,9 +146,7 @@ class Labels(Classes):
     Args:
         labels: A list of labels, assumed to map the class index to the label for that class. If ``labels`` is not
             provided, will attempt to get them from the :class:`.LabelsState`.
-
         multi_label: If true, treats outputs as multi label logits.
-
         threshold: The threshold to use for multi_label classification.
     """
 
@@ -153,8 +158,6 @@ class Labels(Classes):
             self.set_state(LabelsState(labels))
 
     def serialize(self, sample: Any) -> Union[int, List[int], str, List[str]]:
-        sample = sample[DefaultDataKeys.PREDS] if isinstance(sample, Dict) else sample
-        sample = torch.tensor(sample)
         labels = None
 
         if self._labels is not None: