+++ flash/core/serve/flash_components.py
@@ -1,13 +1,13 @@
 import inspect
-from pathlib import Path
-from typing import Any, Callable, Optional, Type
+from typing import Any, Callable, Mapping, Optional
 
 import torch
 from pytorch_lightning.trainer.states import RunningStage
 
 from flash import Task
-from flash.core.serve import Composition, expose, GridModel, ModelComponent
-from flash.core.serve.core import FilePath, GridModelValidArgs_T, GridserveScriptLoader
+from flash.core.data.data_source import DefaultDataKeys
+from flash.core.serve import expose, ModelComponent
+from flash.core.serve.core import FilePath, GridserveScriptLoader
 from flash.core.serve.types.base import BaseType
 
 
@@ -34,9 +34,17 @@ class FlashOutputs(BaseType):
     ):
         self._serializer = serializer
 
-    def serialize(self, output) -> Any:  # pragma: no cover
-        result = self._serializer(output)
-        return result
+    def serialize(self, outputs) -> Any:  # pragma: no cover
+        results = []
+        if isinstance(outputs, list) or isinstance(outputs, torch.Tensor):
+            for output in outputs:
+                result = self._serializer(output)
+                if isinstance(result, Mapping):
+                    result = result[DefaultDataKeys.PREDS]
+                results.append(result)
+        if len(results) == 1:
+            return results[0]
+        return results
 
     def deserialize(self, data: str) -> Any:  # pragma: no cover
         return None
@@ -49,3 +57,39 @@ class FlashServeScriptLoader(GridserveScriptLoader):
     def __init__(self, location: FilePath):
         self.location = location
         self.instance = self.model_cls.load_from_checkpoint(location)
+
+
+def build_flash_serve_model_component(model):
+
+    data_pipeline = model.build_data_pipeline()
+
+    class FlashServeModelComponent(ModelComponent):
+
+        def __init__(self, model):
+            self.model = model
+            self.model.eval()
+            self.data_pipeline = model.build_data_pipeline()
+            self.worker_preprocessor = self.data_pipeline.worker_preprocessor(RunningStage.PREDICTING, is_serving=True)
+            self.device_preprocessor = self.data_pipeline.device_preprocessor(RunningStage.PREDICTING)
+            self.postprocessor = self.data_pipeline.postprocessor(RunningStage.PREDICTING, is_serving=True)
+            # todo (tchaton) Remove this hack
+            self.extra_arguments = len(inspect.signature(self.model.transfer_batch_to_device).parameters) == 3
+            self.device = self.model.device
+
+        @expose(
+            inputs={"inputs": FlashInputs(data_pipeline.deserialize_processor())},
+            outputs={"outputs": FlashOutputs(data_pipeline.serialize_processor())},
+        )
+        def predict(self, inputs):
+            with torch.no_grad():
+                inputs = self.worker_preprocessor(inputs)
+                if self.extra_arguments:
+                    inputs = self.model.transfer_batch_to_device(inputs, self.device, 0)
+                else:
+                    inputs = self.model.transfer_batch_to_device(inputs, self.device)
+                inputs = self.device_preprocessor(inputs)
+                preds = self.model.predict_step(inputs, 0)
+                preds = self.postprocessor(preds)
+                return preds
+
+    return FlashServeModelComponent(model)