+++ tests/image/embedding/test_model.py
@@ -16,11 +16,11 @@ import os
 import pytest
 import torch
 
-from flash.core.utilities.imports import _IMAGE_AVAILABLE
 from flash.image import ImageEmbedder
+from tests.helpers.utils import _IMAGE_TESTING
 
 
-@pytest.mark.skipif(not _IMAGE_AVAILABLE, reason="image libraries aren't installed.")
+@pytest.mark.skipif(not _IMAGE_TESTING, reason="image libraries aren't installed.")
 @pytest.mark.parametrize("jitter, args", [(torch.jit.script, ()), (torch.jit.trace, (torch.rand(1, 3, 32, 32), ))])
 def test_jit(tmpdir, jitter, args):
     path = os.path.join(tmpdir, "test.pt")