+++ tests/tabular/classification/test_data_model_integration.py
@@ -16,6 +16,7 @@ import pytorch_lightning as pl
 
 from flash.core.utilities.imports import _TABULAR_AVAILABLE
 from flash.tabular import TabularClassifier, TabularData
+from tests.helpers.utils import _TABULAR_TESTING
 
 if _TABULAR_AVAILABLE:
     import pandas as pd
@@ -30,7 +31,7 @@ if _TABULAR_AVAILABLE:
     )
 
 
-@pytest.mark.skipif(not _TABULAR_AVAILABLE, reason="tabular libraries aren't installed.")
+@pytest.mark.skipif(not _TABULAR_TESTING, reason="tabular libraries aren't installed.")
 def test_classification(tmpdir):
 
     train_data_frame = TEST_DF_1.copy()