+++ tests/text/test_data_model_integration.py
@@ -17,8 +17,8 @@ from pathlib import Path
 import pytest
 from pytorch_lightning import Trainer
 
-from flash.core.utilities.imports import _TEXT_AVAILABLE
 from flash.text import TextClassificationData, TextClassifier
+from tests.helpers.utils import _TEXT_TESTING
 
 TEST_BACKBONE = "prajjwal1/bert-tiny"  # super small model for testing
 
@@ -36,7 +36,7 @@ def csv_data(tmpdir):
 
 
 @pytest.mark.skipif(os.name == "nt", reason="Huggingface timing out on Windows")
-@pytest.mark.skipif(not _TEXT_AVAILABLE, reason="text libraries aren't installed.")
+@pytest.mark.skipif(not _TEXT_TESTING, reason="text libraries aren't installed.")
 def test_classification(tmpdir):
 
     csv_path = csv_data(tmpdir)